+++ src/DomNodeListAction/DomNodeListActionInterface.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Xparse\ElementFinder\DomNodeListAction;
+
+interface DomNodeListActionInterface
+{
+    public function execute(\DOMNodeList $nodeList): void;
+}